+++ FoodTrucker.Models/Ingredient/IngredientCreate.cs
@@ -15,7 +15,7 @@ namespace FoodTrucker.Models.Customer
         public string Name { get; set; }
         [Required]
         [Range(0, 150)]
-        public double QuantityInStock { get; set; }
+        public int QuantityInStock { get; set; }
         [Required]
         public IngredientType IngredientType { get; set; }
     }
